move near_network_primitives::metrics back into near_network
