+++ chain/network-primitives/src/lib.rs
@@ -1,6 +1,5 @@
 #[cfg(feature = "metric_recorder")]
 pub mod recorder;
 
-pub mod metrics;
 pub mod routing;
 pub mod types;
